# docs/swagger.yaml
...
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  schema.GetPrivilegesConfigResp:
    properties:
      options:
        items:
          $ref: '#/definitions/schema.PrivilegeOption'
        type: array
      selected_level:
        $ref: '#/definitions/schema.PrivilegeLevel'
    type: object
...
  schema.GetSMTPConfigResp:
    properties:
      encryption:
        description: '"" SSL'
        type: string
      from_email:
        type: string
      from_name:
        type: string
      smtp_authentication:
        type: boolean
      smtp_host:
        type: string
      smtp_password:
        type: string
      smtp_port:
        type: integer
      smtp_username:
        type: string
    type: object
...
  schema.PrivilegeLevel:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - PrivilegeLevel1
    - PrivilegeLevel2
    - PrivilegeLevel3
...
  schema.SiteBrandingReq:
    properties:
      favicon:
        maxLength: 512
        type: string
      logo:
        maxLength: 512
        type: string
      mobile_logo:
        maxLength: 512
        type: string
      square_icon:
        maxLength: 512
        type: string
    type: object
  schema.SiteBrandingResp:
    properties:
      favicon:
        maxLength: 512
        type: string
      logo:
        maxLength: 512
        type: string
      mobile_logo:
        maxLength: 512
        type: string
      square_icon:
        maxLength: 512
        type: string
    type: object
  schema.SiteCustomCssHTMLReq:
    properties:
      custom_css:
        maxLength: 65536
        type: string
      custom_footer:
        maxLength: 65536
        type: string
      custom_head:
        maxLength: 65536
        type: string
      custom_header:
        maxLength: 65536
        type: string
      custom_sidebar:
        maxLength: 65536
        type: string
    type: object
  schema.SiteCustomCssHTMLResp:
    properties:
      custom_css:
        maxLength: 65536
        type: string
      custom_footer:
        maxLength: 65536
        type: string
      custom_head:
        maxLength: 65536
        type: string
      custom_header:
        maxLength: 65536
        type: string
      custom_sidebar:
        maxLength: 65536
        type: string
    type: object
  schema.SiteGeneralReq:
    properties:
      contact_email:
        maxLength: 512
        type: string
      description:
        maxLength: 2000
        type: string
      name:
        maxLength: 128
        type: string
      short_description:
        maxLength: 255
        type: string
      site_url:
        maxLength: 512
        type: string
    required:
    - contact_email
    - name
    - site_url
    type: object
  schema.SiteGeneralResp:
    properties:
      contact_email:
        maxLength: 512
        type: string
      description:
        maxLength: 2000
        type: string
      name:
        maxLength: 128
        type: string
      short_description:
        maxLength: 255
        type: string
      site_url:
        maxLength: 512
        type: string
    required:
    - contact_email
    - name
    - site_url
    type: object
...
  schema.SiteInterfaceReq:
    properties:
      language:
        maxLength: 128
        type: string
      time_zone:
        maxLength: 128
        type: string
    required:
    - language
    - time_zone
    type: object
  schema.SiteInterfaceResp:
    properties:
      language:
        maxLength: 128
        type: string
      time_zone:
        maxLength: 128
        type: string
    required:
    - language
    - time_zone
    type: object
  schema.SiteLegalReq:
    properties:
      privacy_policy_original_text:
        type: string
      privacy_policy_parsed_text:
        type: string
      terms_of_service_original_text:
        type: string
      terms_of_service_parsed_text:
        type: string
    type: object
  schema.SiteLegalResp:
    properties:
      privacy_policy_original_text:
        type: string
      privacy_policy_parsed_text:
        type: string
      terms_of_service_original_text:
        type: string
      terms_of_service_parsed_text:
        type: string
    type: object
  schema.SiteLoginReq:
    properties:
      allow_email_domains:
        items:
          type: string
        type: array
      allow_email_registrations:
        type: boolean
      allow_new_registrations:
        type: boolean
      login_required:
        type: boolean
    type: object
  schema.SiteLoginResp:
    properties:
      allow_email_domains:
        items:
          type: string
        type: array
      allow_email_registrations:
        type: boolean
      allow_new_registrations:
        type: boolean
      login_required:
        type: boolean
    type: object
  schema.SiteSeoReq:
    properties:
      permalink:
        maximum: 4
        minimum: 0
        type: integer
      robots:
        type: string
    required:
    - permalink
    - robots
    type: object
  schema.SiteSeoResp:
    properties:
      permalink:
        maximum: 4
        minimum: 0
        type: integer
      robots:
        type: string
    required:
    - permalink
    - robots
    type: object
  schema.SiteThemeReq:
    properties:
      theme:
        maxLength: 255
        type: string
      theme_config:
        additionalProperties: true
        type: object
    required:
    - theme
    type: object
  schema.SiteThemeResp:
    properties:
      theme:
        type: string
      theme_config:
        additionalProperties: true
        type: object
      theme_options:
        items:
          $ref: '#/definitions/schema.ThemeOption'
        type: array
    type: object
  schema.SiteUsersReq:
    properties:
      allow_update_avatar:
        type: boolean
      allow_update_bio:
        type: boolean
      allow_update_display_name:
        type: boolean
      allow_update_location:
        type: boolean
      allow_update_username:
        type: boolean
      allow_update_website:
        type: boolean
      default_avatar:
        enum:
        - system
        - gravatar
        type: string
      gravatar_base_url:
        type: string
    required:
    - default_avatar
    type: object
  schema.SiteUsersResp:
    properties:
      allow_update_avatar:
        type: boolean
      allow_update_bio:
        type: boolean
      allow_update_display_name:
        type: boolean
      allow_update_location:
        type: boolean
      allow_update_username:
        type: boolean
      allow_update_website:
        type: boolean
      default_avatar:
        enum:
        - system
        - gravatar
        type: string
      gravatar_base_url:
        type: string
    required:
    - default_avatar
    type: object
  schema.SiteWriteReq:
    properties:
      recommend_tags:
        items:
          type: string
        type: array
      required_tag:
        type: boolean
      reserved_tags:
        items:
          type: string
        type: array
    type: object
  schema.SiteWriteResp:
    properties:
      recommend_tags:
        items:
          type: string
        type: array
      required_tag:
        type: boolean
      reserved_tags:
        items:
          type: string
        type: array
    type: object
...
  schema.UpdatePrivilegesConfigReq:
    properties:
      level:
        allOf:
        - $ref: '#/definitions/schema.PrivilegeLevel'
        maximum: 3
        minimum: 1
    required:
    - level
    type: object
  schema.UpdateSMTPConfigReq:
    properties:
      encryption:
        description: '"" SSL'
        enum:
        - SSL
        type: string
      from_email:
        maxLength: 256
        type: string
      from_name:
        maxLength: 256
        type: string
      smtp_authentication:
        type: boolean
      smtp_host:
        maxLength: 256
        type: string
      smtp_password:
        maxLength: 256
        type: string
      smtp_port:
        maximum: 65535
        minimum: 1
        type: integer
      smtp_username:
        maxLength: 256
        type: string
      test_email_recipient:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /answer/admin/api/setting/privileges:
    get:
      description: GetPrivilegesConfig get privileges config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetPrivilegesConfigResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: GetPrivilegesConfig get privileges config
      tags:
      - admin
    put:
      description: update privileges config
      parameters:
      - description: config
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.UpdatePrivilegesConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update privileges config
      tags:
      - admin
  /answer/admin/api/setting/smtp:
    get:
      description: GetSMTPConfig get smtp config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetSMTPConfigResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: GetSMTPConfig get smtp config
      tags:
      - admin
    put:
      description: update smtp config
      parameters:
      - description: smtp config
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.UpdateSMTPConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update smtp config
      tags:
      - admin
  /answer/admin/api/siteinfo/branding:
    get:
      description: get site interface
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.SiteBrandingResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: get site interface
      tags:
      - admin
    put:
      description: update site info branding
      parameters:
      - description: branding info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SiteBrandingReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update site info branding
      tags:
      - admin
  /answer/admin/api/siteinfo/custom-css-html:
    get:
      description: get site info custom html css config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.SiteCustomCssHTMLResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: get site info custom html css config
      tags:
      - admin
    put:
      description: update site custom css html config
      parameters:
      - description: login info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SiteCustomCssHTMLReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update site custom css html config
      tags:
      - admin
  /answer/admin/api/siteinfo/general:
    get:
      description: get site general information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.SiteGeneralResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: get site general information
      tags:
      - admin
    put:
      description: update site general information
      parameters:
      - description: general
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SiteGeneralReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update site general information
      tags:
      - admin
  /answer/admin/api/siteinfo/interface:
    get:
      description: get site interface
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.SiteInterfaceResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: get site interface
      tags:
      - admin
    put:
      description: update site info interface
      parameters:
      - description: general
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SiteInterfaceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update site info interface
      tags:
      - admin
  /answer/admin/api/siteinfo/legal:
    get:
      description: Set the legal information for the site
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.SiteLegalResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Set the legal information for the site
      tags:
      - admin
    put:
      description: update site legal info
      parameters:
      - description: write info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SiteLegalReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update site legal info
      tags:
      - admin
  /answer/admin/api/siteinfo/login:
    get:
      description: get site info login config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.SiteLoginResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: get site info login config
      tags:
      - admin
    put:
      description: update site login
      parameters:
      - description: login info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SiteLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update site login
      tags:
      - admin
  /answer/admin/api/siteinfo/seo:
    get:
      description: get site seo information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.SiteSeoResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: get site seo information
      tags:
      - admin
    put:
      description: update site seo information
      parameters:
      - description: seo
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SiteSeoReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update site seo information
      tags:
      - admin
  /answer/admin/api/siteinfo/theme:
    get:
      description: get site info theme config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.SiteThemeResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: get site info theme config
      tags:
      - admin
    put:
      description: update site custom css html config
      parameters:
      - description: login info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SiteThemeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update site custom css html config
      tags:
      - admin
  /answer/admin/api/siteinfo/users:
    get:
      description: get site user config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.SiteUsersResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: get site user config
      tags:
      - admin
    put:
      description: update site info config about users
      parameters:
      - description: users info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SiteUsersReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update site info config about users
      tags:
      - admin
  /answer/admin/api/siteinfo/write:
    get:
      description: get site interface
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.SiteWriteResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: get site interface
      tags:
      - admin
    put:
      description: update site write info
      parameters:
      - description: write info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SiteWriteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update site write info
      tags:
      - admin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"